terraform-applications/brewery/functions/beer_demo_data_pipeline/requirements.txt
#google-cloud-firestore
#google-cloud-pubsub
google-cloud-bigquery
#opentelemetry-exporter-google-cloud